requirements.txt
fastapi
uvicorn
pypdf
openai
python-dotenv
python-multipart
python-docx
pdf2image
pytesseract
pillow
pydantic




